Add OpenAPI specification

+++ backend/GalaxyBoardsAPI/Controllers/ProjectsController.cs
@@ -9,6 +9,7 @@ namespace GalaxyBoardsAPI.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
+    [Produces("application/json")]
     public class ProjectsController : ControllerBase
     {
         private readonly IEntityRepository<Project> _projectRepo;
@@ -22,7 +23,9 @@ namespace GalaxyBoardsAPI.Controllers
             _ticketRepo = ticketRepo;
         }
 
-        // GET: api/projects
+        /// <summary>
+        /// Lists stored projects
+        /// </summary>
         [HttpGet]
         public ActionResult<QueryResult<ProjectBrief>> QueryProjects()
         {
@@ -33,7 +36,11 @@ namespace GalaxyBoardsAPI.Controllers
             };
         }
 
-        // GET: api/projects/5
+        /// <summary>
+        /// Looks up a project by ID
+        /// </summary>
+        /// <param name="id">ID of project to find</param>
+        /// <response code="200">Project found matching ID</response>
         [HttpGet("{id}")]
         public ActionResult<ProjectDetail> GetProject(Guid id)
         {
@@ -47,7 +54,13 @@ namespace GalaxyBoardsAPI.Controllers
             return projectDto;
         }
 
-        // PUT: api/projects/5
+        /// <summary>
+        /// Updates an existing project
+        /// </summary>
+        /// <param name="id">ID of the project to update</param>
+        /// <param name="updated">Project update data</param>
+        /// <response code="200">Project updated successfully</response>
+        /// <response code="404">Project not found</response>
         [HttpPut("{id}")]
         public IActionResult PutProject(Guid id, ProjectUpdate updated)
         {
@@ -63,7 +76,11 @@ namespace GalaxyBoardsAPI.Controllers
             return Ok();
         }
 
-        // POST: api/projects
+        /// <summary>
+        /// Creates a new project
+        /// </summary>
+        /// <param name="project">Project data</param>
+        /// <response code="200">Project successfully created</response>
         [HttpPost]
         public ActionResult<PostResult> PostProject(ProjectPostData project)
         {
@@ -84,17 +101,27 @@ namespace GalaxyBoardsAPI.Controllers
             };
         }
 
-        // DELETE: api/projects/5
+        /// <summary>
+        /// Deletes a project and all associated boards and tickets
+        /// </summary>
+        /// <param name="id">ID of project to delete</param>
+        /// <response code="200">Project successfully deleted</response>
+        /// <response code="404">Project not found</response>
         [HttpDelete("{id}")]
         public IActionResult DeleteProject(Guid id)
         {
-            var projectTickets = _ticketRepo.Get(filter: t => t.Project.Id == id).ToList();
-            projectTickets.ForEach(t =>
+            var project = _projectRepo.GetById(id);
+            if (project == null)
+            {
+                return NotFound();
+            }
+
+            project.Tickets.ToList().ForEach(t =>
             {
                 _ticketRepo.Delete(t.Id);
             });
+            
             _projectRepo.Delete(id);
-            _ticketRepo.Save();
             _projectRepo.Save();
             return Ok();
         }
